==> src/notebooks/4-home-credit-submit.ipynb <==
from collections import OrderedDict
import gc
import pathlib
from pprint import pprint
import random
from typing import Dict, List, Tuple, Union

from catboost import CatBoostClassifier
from joblib import load
import lightgbm as lgb
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import polars as pl
import seaborn as sns
from sklearn.linear_model import LinearRegression
from sklearn.metrics import roc_auc_score
from sklearn.preprocessing import OrdinalEncoder


gc.enable()

competition_dir = pathlib.Path('/kaggle/input/home-credit-credit-risk-model-stability')
feature_dtype_dir = pathlib.Path('/kaggle/input/0-home-credit-make-dtype-csv')
model_dir = pathlib.Path('/kaggle/input/home-credit-04251529')

bool_features = pd.read_csv(feature_dtype_dir.joinpath('bool_features.csv'))
float64_features = pd.read_csv(feature_dtype_dir.joinpath('float64_features.csv'))
string_features = pd.read_csv(feature_dtype_dir.joinpath('string_features.csv'))
date_features = pd.read_csv(feature_dtype_dir.joinpath('date_features.csv'))

bool_features['cast_dtype'] = pl.Int8
float64_features['cast_dtype'] = pl.Float32
string_features['cast_dtype'] = pl.String
date_features['cast_dtype'] = pl.Date

from collections import OrderedDict
import pathlib


def get_depth_paths(load_dir: pathlib.Path, prefix: str):
    
    assert prefix in ['test', 'train']
    
    depth_paths = OrderedDict()

    depth_paths['static_0'] = []
    depth_paths['static_cb_0'] = []
    depth_paths['applprev_1'] = []
    depth_paths['other_1'] = []
    depth_paths['tax_registry_a_1'] = []
    depth_paths['tax_registry_b_1'] = []
    depth_paths['tax_registry_c_1'] = []
    depth_paths['credit_bureau_a_1'] = []
    depth_paths['credit_bureau_b_1'] = []
    depth_paths['deposit_1'] = []
    depth_paths['person_1'] = []
    depth_paths['debitcard_1'] = []
    depth_paths['applprev_2'] = []
    depth_paths['person_2'] = []
    depth_paths['credit_bureau_a_2'] = []
    depth_paths['credit_bureau_b_2'] = []

    for k in depth_paths.keys():
        depth_paths[k] = sorted(
            [p for p in load_dir.joinpath(f'{prefix}').glob(f'{prefix}_{k}*.parquet')]
        )
    return depth_paths

test_depth_paths = get_depth_paths(competition_dir.joinpath('parquet_files'), 'test')
print(f'the number of test paths: {sum(len(v1) for v1 in test_depth_paths.values())}')

from datetime import datetime
from typing import List

import polars as pl
import polars.selectors as cs


def process_date_features(data: pl.DataFrame) -> pl.DataFrame:
    data = data.with_columns(
        pl.col(pl.Date).dt.year().cast(pl.Float32).name.prefix('year_'),
        pl.col(pl.Date).dt.month().cast(pl.Float32).name.prefix('month_'),
        pl.col(pl.Date).dt.day().cast(pl.Float32).name.prefix('day_'),
    )
    data = data.drop(cs.date())
    return data

import gc
from itertools import combinations
import pathlib
from typing import Dict, List

import numpy as np
import pandas as pd
import polars as pl


def merge_dataset(
        base_data: pl.DataFrame,
        depth_paths: Dict[str, List[pathlib.Path]],
        bool_features: pd.DataFrame,
        float64_features: pd.DataFrame,
        string_features: pd.DataFrame,
        date_features: pd.DataFrame,
        useful_features: pd.DataFrame = None,
        depth: str = '012',
    ) -> pd.DataFrame:
    
    assert depth in ['0', '1', '2', '012']

    aggs = [
        pl.col(pl.Float32).max().cast(pl.Float32).name.prefix('max_'),
        pl.col(pl.Float32).median().cast(pl.Float32).name.prefix('median_'),
        pl.col(pl.Float32).sum().cast(pl.Float32).name.prefix('sum_'),
        pl.col(pl.Int32, pl.Int64).max().cast(pl.Int32).name.prefix('max_'),
        pl.col(pl.Int32, pl.Int64).median().cast(pl.Int32).name.prefix('median_'),
        pl.col(pl.Int32, pl.Int64).sum().cast(pl.Int32).name.prefix('sum_'),
        pl.col(pl.Date, pl.String).first().name.prefix('first_'),
    ]
    
    for i, (k, path_list) in enumerate(depth_paths.items()):
        
        if depth == '012':
            pass
        elif depth != k[-1]:
            continue
            
        print(f'loading `{k}`')
        depth_data = []
        for p in path_list:
            sub_data = pl.read_parquet(p).cast({'case_id': pl.Int64})

            '''
            cast dtypes
            '''
            for _, row in bool_features[['Variable', 'cast_dtype']].iterrows():
                col = row['Variable']
                cast_dtype = row['cast_dtype']
                if col in sub_data.columns:
                    sub_data = sub_data.with_columns(pl.col(col).fill_null(2).fill_nan(2).cast(cast_dtype))
                
            for _, row in float64_features[['Variable', 'cast_dtype']].iterrows():
                col = row['Variable']
                cast_dtype = row['cast_dtype']
                if col in sub_data.columns:
                    sub_data = sub_data.with_columns(pl.col(col).cast(cast_dtype))
                    
            for _, row in string_features[['Variable', 'cast_dtype']].iterrows():
                col = row['Variable']
                cast_dtype = row['cast_dtype']
                if col in sub_data.columns:
                    sub_data = sub_data.with_columns(pl.col(col).fill_null('none').cast(cast_dtype))

            for _, row in date_features[['Variable', 'cast_dtype']].iterrows():
                col = row['Variable']
                cast_dtype = row['cast_dtype']
                if col in sub_data.columns:
                    sub_data = sub_data.with_columns(pl.col(col).cast(cast_dtype))

            '''
            rename columns
            '''
            sub_data = sub_data.rename(lambda c: rename_column(c, k[-1]))
            
            '''
            delete useless features
            '''
            if useful_features is not None:
                sub_data = sub_data.drop(
                    [
                        col
                        for col in sub_data.columns
                        if (
                            (col not in useful_features['Variable'].to_list())
                            and
                            (col != 'case_id' and col != 'date_decision' and col != 'MONTH' and col != 'WEEK_NUM')
                        )
                    ]
                )
            
            '''
            aggregation
            '''
            if k[-1] == '1' :
                aggs_12 = aggs + [
                    (pl.col(col).first() / pl.col(col).max()).cast(pl.Float32).name.prefix('ratio_')
                    for col in sub_data.select(pl.col(pl.Float32)).columns
                ]
                
                sub_data = sub_data.group_by('case_id').agg(aggs_12).sort('case_id')
                # for col in sub_data.columns:
                #     if col.startswith('head'):
                #         sub_data = sub_data.with_columns(
                #             pl.col(col).list.get(0).name.prefix('first_'),
                #             pl.col(col).list.get(1).name.prefix('second_'),
                #         )
                #         sub_data = sub_data.drop(col)
            elif k[-1] == '2':
                aggs_12 = aggs + [
                    (pl.col(col).first() / pl.col(col).max()).cast(pl.Float32).name.prefix('ratio_')
                    for col in sub_data.select(pl.col(pl.Float32)).columns
                ]
                sub_data = sub_data.group_by(['case_id', 'num_group1']).agg(aggs_12).group_by('case_id').agg(aggs).sort('case_id')
            sub_data = sub_data.drop([col for col in sub_data.columns if 'num_group' in col])

            depth_data.append(sub_data)
            print(f'\t{sub_data.shape}')
            
            del sub_data
            gc.collect()
        
        depth_data = pl.concat(depth_data, how='vertical_relaxed')
        base_data = base_data.join(depth_data, how='left', on='case_id', suffix=f'_{i}')
        
        del depth_data
        gc.collect()

    # '''
    # add new features
    # '''
    # depth_0_P_high_fimp_features = [
    #     'avgdpdtolclosure24_3658938P_0',
    #     'maxdbddpdtollast12m_3658940P_0',
    #     'maxdpdlast3m_392P_0',
    # ]

    # depth_0_A_high_fimp_features = [
    #     'price_1097A_0',
    #     'pmtssum_45A_0',
    #     'annuity_780A_0',
    #     'credamount_770A_0',
    # ]

    # depth_0_L_high_fimp_features = [
    #     'pmtnum_254L_0',
    #     'mobilephncnt_593L_0',
    #     'days180_256L_0',
    #     'days120_123L_0',
    #     'eir_270L_0',
    #     'numrejects9m_859L_0',
    #     'isbidproduct_1095L_0',
    #     'days90_310L_0',
    #     'days360_512L_0',
    #     'pctinstlsallpaidlate1d_3546856L_0',
    #     'numinstpaidearly3d_3546850L_0',
    #     'pmtscount_423L_0',
    #     'numinstunpaidmax_3546851L_0',
    #     'cntpmts24_3658933L_0',
    #     'numinstlsallpaid_934L_0',
    # ]

    # display(base_data.select(depth_0_L_high_fimp_features))

    # P_aggs = []
    # for col1, col2 in combinations(depth_0_P_high_fimp_features, 2):
    #     P_aggs.append((pl.col(col1) - pl.col(col2)).alias(f'diff_{col1}_{col2}'))
    #     P_aggs.append((pl.col(col1) + pl.col(col2)).alias(f'sum_{col1}_{col2}'))

    # A_aggs = []
    # for col1, col2 in combinations(depth_0_A_high_fimp_features, 2):
    #     A_aggs.append((pl.col(col1) - pl.col(col2)).alias(f'diff_{col1}_{col2}'))
    #     A_aggs.append((pl.col(col1) + pl.col(col2)).alias(f'sum_{col1}_{col2}'))

    # L_aggs = []
    # for col1, col2 in combinations(depth_0_L_high_fimp_features, 2):
    #     L_aggs.append((pl.col(col1) - pl.col(col2)).alias(f'diff_{col1}_{col2}'))
    #     L_aggs.append((pl.col(col1) + pl.col(col2)).alias(f'sum_{col1}_{col2}'))

    # base_data = base_data.with_columns(*A_aggs)

    '''
    process date features
    '''
    base_data = process_date_features(base_data)

    '''
    convert polars DataFrame into pandas DataFrame
    '''
    base_data = base_data.to_pandas()

    return base_data


def rename_column(column, depth):
    if column in ['case_id', 'num_group1', 'num_group2']:
        return column
    else:
        return column + f'_{depth}'

test_base_data = pl.read_parquet(
    competition_dir.joinpath('parquet_files/test/test_base.parquet')
)
test_base_data = test_base_data.cast(
    {
        'case_id': pl.Int64,
        'date_decision': pl.String,
        'MONTH': pl.Int64,
        'WEEK_NUM': pl.Int64,
    }
)
display(test_base_data)

%%time


depth_data = merge_dataset(
    test_base_data,
    test_depth_paths,
    bool_features,
    float64_features,
    string_features,
    date_features,
    None,
    '012',
)
display(depth_data)

models = []
for k in range(5):
    #models.append(lgb.Booster(model_file=model_dir.joinpath(f'lgb_fold{k+1}_04250706.txt')))
    models.append(CatBoostClassifier().load_model(model_dir.joinpath(f'catboost_fold{k+1}_04251529.cbm')))
    
training_features = np.load(model_dir.joinpath('training_features.npy'), allow_pickle=True)
depth_data = depth_data[training_features]
display(depth_data)

# def encode_objects(depth_data: pd.DataFrame) -> pd.DataFrame:
#     object_columns = depth_data.dtypes.index[depth_data.dtypes==object].to_list()
#     if 'date_decision' in object_columns:
#         object_columns.remove('date_decision')
    
#     object_data = []
#     for col in object_columns:
#         encoder = load(model_dir.joinpath(f'encoder_{col}.joblib'))
#         object_data.append(encoder.transform(depth_data[col].values.reshape(-1, 1)).astype(np.float32))
#     depth_data.drop(columns=object_columns, inplace=True)
#     object_data = np.concatenate(object_data, axis=1)
#     object_data = pd.DataFrame(object_data, columns=object_columns)
    
#     return pd.concat([depth_data, object_data], axis=1)
    
# display(depth_data)
# depth_data = encode_objects(depth_data)
# display(depth_data)

def predict(
        models: List[object],
        submit: pd.DataFrame,
        X: pd.DataFrame,
        batch_size: int = 1024,
    ) -> pd.DataFrame:
    
    categorical_features = X.dtypes.index[X.dtypes==object].to_list()
    X.fillna({col: 'none' for col in categorical_features}, inplace=True)
    
    probas = np.zeros(len(X), dtype=np.float32)
    for i in range(0, len(X), batch_size):
        limit = i + batch_size if i + batch_size < len(X) else len(X)
        batch_X = X.iloc[i:limit]
        
        for model in models:
            # proba = model.predict(batch_X, num_iteration=model.best_iteration)
            proba = model.predict_proba(batch_X)[:, 1]
            probas[i:limit] += proba.astype(np.float32)
            
            del proba
            gc.collect()
            
        del batch_X
        gc.collect()
    
    probas /= len(models)
    
    submit['score'] = probas
    
    return submit

submission = predict(models, test_base_data.to_pandas(), depth_data, 2048)
display(submission)

submission = submission[['case_id', 'score']]
display(submission)

# 

submission.to_csv('submission.csv', index=False)
